# smartender-global/smartender-global-backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Description of the global smartender API"
  version: "1.0.0"
  title: "smartender-global-server"
host: "vahlkamp.selfhost.eu:10000"
basePath: "/smartender/api"
...
schemes:
- "https"
- "http"
security:
- password: []
paths:
...
  /sessions/current:
    get:
      tags:
      - "administration"
      operationId: "sessionsCurrentGET"
      parameters: []
      responses:
        200:
          description: "the active session"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Session"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/join:
    post:
      tags:
      - "administration"
      description: "join a session. A pending invite is required for that"
      operationId: "sessionsJoinPOST"
      parameters:
      - name: "guid"
        in: "query"
        description: "global unique identifier of the session to join"
        required: true
        type: "string"
      responses:
        202:
          description: "current user joined the session"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/leave:
    post:
      tags:
      - "administration"
      description: "leave a session. You have to be part of the session to leave"
      operationId: "sessionsLeavePOST"
      parameters:
      - name: "guid"
        in: "query"
        description: "global unique identifier of the session to leave"
        required: true
        type: "string"
      responses:
        202:
          description: "current user left the session"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/invite:
    post:
      tags:
      - "administration"
      description: "invite a user to a session"
      operationId: "sessionsInvitePOST"
      parameters:
      - name: "userid"
        in: "query"
        required: true
        type: "integer"
      - name: "sessionid"
        in: "query"
        required: true
        type: "integer"
      - name: "canEditMachine"
        in: "query"
        description: "if set to true, the user will be able to perform operations\
          \ like setting slot values or ingredients"
        required: false
        type: "boolean"
        default: false
      responses:
        202:
          description: "user was invited to the session"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/kick:
    post:
      tags:
      - "administration"
      description: "kicks a user from a session"
      operationId: "sessionsKickPOST"
      parameters:
      - name: "userid"
        in: "query"
        required: true
        type: "integer"
      - name: "sessionid"
        in: "query"
        required: true
        type: "integer"
      responses:
        202:
          description: "user was kicked from the session"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/create:
    post:
      tags:
      - "administration"
      description: "creates a new session for the logged-in user and the specified\
        \ machine"
      operationId: "sessionsCreatePOST"
      parameters:
      - name: "machine"
        in: "query"
        description: "id of the machine for this session"
        required: true
        type: "integer"
      responses:
        201:
          description: "created a nes session and returned session object"
          schema:
            $ref: "#/definitions/Session"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/delete:
    delete:
      tags:
      - "administration"
      description: "deletes the specified session. Logged-in user must be the owner"
      operationId: "sessionsDeleteDELETE"
      parameters:
      - name: "session"
        in: "query"
        description: "guid of the session to delete"
        required: true
        type: "string"
      responses:
        201:
          description: "session successfully deleted"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /sessions/pause:
    post:
      tags:
      - "administration"
      description: "deactivates a session"
      operationId: "sessionsPausePOST"
      parameters:
      - name: "sessionid"
        in: "query"
        description: "guid of the session to pause"
        required: true
        type: "string"
      responses:
        201:
          description: "session successfully deleted"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
...
  /machines/associate:
    post:
      tags:
      - "administration"
      operationId: "machinesAssociatePOST"
      parameters:
      - in: "body"
        name: "machinekey"
        description: "unique key associated with the machine"
        required: true
        schema:
          type: "string"
      responses:
        202:
          description: "user successfully associated with the machine"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
    delete:
      tags:
      - "administration"
      operationId: "machinesAssociateDELETE"
      parameters:
      - in: "body"
        name: "machinekey"
        description: "unique key associated with the machine"
        required: true
        schema:
          type: "string"
      responses:
        202:
          description: "machine successfully removed from user"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /machines/{machineid}/fillslot:
    post:
      tags:
      - "administration"
      operationId: "machinesMachineidFillslotPOST"
      parameters:
      - name: "machineid"
        in: "path"
        description: "the id of the machine htat this operation will be executed for"
        required: true
        type: "integer"
      - name: "ingredientid"
        in: "query"
        description: "id of the ingredient that ill be put in the slot"
        required: true
        type: "integer"
      responses:
        202:
          description: "successfully filled slot"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /machines/{machineid}/clearslot:
    post:
      tags:
      - "administration"
      operationId: "machinesMachineidClearslotPOST"
      parameters:
      - name: "machineid"
        in: "path"
        description: "the id of the machine htat this operation will be executed for"
        required: true
        type: "integer"
      responses:
        202:
          description: "successfully cleared slot"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /machines/{machineid}/setslot-state:
    post:
      tags:
      - "administration"
      operationId: "machinesMachineidSetslot_statePOST"
      parameters:
      - name: "machineid"
        in: "path"
        description: "the id of the machine htat this operation will be executed for"
        required: true
        type: "integer"
      - name: "state"
        in: "query"
        description: "the state the slot should be set to"
        required: true
        type: "boolean"
      responses:
        202:
          description: "successfully set slot-state"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
  /machines/{machineid}/close-all:
    post:
      tags:
      - "administration"
      operationId: "machinesMachineidClose_allPOST"
      parameters:
      - name: "machineid"
        in: "path"
        description: "the id of the machine htat this operation will be executed for"
        required: true
        type: "integer"
      responses:
        202:
          description: "successfully closed all slots"
        401:
          description: "Required authorization was missing"
          schema:
            $ref: "#/definitions/inline_response_401"
        403:
          description: "the loggen-in user did not have the required permissions for\
            \ the operation"
          schema:
            $ref: "#/definitions/inline_response_default"
        default:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_default"
      security:
      - password: []
      x-swagger-router-controller: "Administration"
securityDefinitions:
  password:
    type: "basic"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      alias:
        type: "string"
        description: "the users display-name"
    example:
      alias: "alias"
      id: 0
...
  Machine:
    type: "object"
    properties:
      name:
        type: "string"
        description: "name of the machine"
      owner:
        $ref: "#/definitions/User"
      id:
        type: "integer"
        description: "unique id of this machine"
    description: "a machine"
    example:
      owner:
        alias: "alias"
        id: 0
      name: "name"
      id: 6
...
  Session:
    type: "object"
    properties:
      guid:
        type: "string"
        description: "a global unique identifier for this session"
      owner:
        $ref: "#/definitions/User"
      machine:
        $ref: "#/definitions/Machine"
      active:
        type: "boolean"
        description: "true if this is the users active session"
    description: "describes a session"
    example:
      owner:
        alias: "alias"
        id: 0
      machine:
        owner:
          alias: "alias"
          id: 0
        name: "name"
        id: 6
      guid: "guid"
      active: true
  inline_response_401:
    properties:
      required_scopes:
        type: "array"
        description: "scopes that are required for this operation"
        items:
          type: "string"
  inline_response_default:
    properties:
      message:
        type: "string"
...